Start rewriting day11 with a view toward parsing arbitrary inputs
Parsing this input is unexpectedly difficult. I'm fairly confident
in the parser, but actually running it produces this error:

```
input-11.txt:1: Unrecognized token `a` found at 25:26
Expected one of "a" or "nothing relevant" for "The first floor contains a promethium generator and a promethium-compatible microchip."
```

Given that the found token `a` is the same as the expected token "a",
it almost seems like LALRPOP itself is having problems.