solve day13
I was a bit nervous about off-by-1 errors, particularly in part 2,
but it seems like I'm finally getting the right intuition for avoiding
those.

It was _really_ convenient that I happened to have a breadth-first-search
navigation function lying around. (Docs say it's A*, but a quick glance
at the source implies that it doesn't use a heuristic function, so
it really shouldn't be called that.)

Actually, this whole exercise felt like a huge validation of having
a generic tile cartography library prepped and ready for AoC. There
are many more than 100 lines of code _involved_, but I didn't have to
write that many _today_.