solve day14 part 2
As so often happens, part 2 took an unreasonably long time.
Ostensibly, the difficulty was in making the stretched hash function,
and ensuring that the runtime didn't explode. That was implemented
and tested pretty quickly, and I was feeling smug.

Then I _just couldn't get the right answer_.

I'd previously had a runtime assertion that no key was both inserted
and completed in the same index pair. I came up with that on my own,
not knowing the shape of the exercise to come. However, it turns out
that the problem specification had already handled it in a different
way: we're only allowed to consider the first triple for any given hash.

It took hours, tests, and going to reddit to notice that only the
first triple was allowed. Once that was resolved, everything came
together very quickly.

Maybe one day I'll learn to notice _all_ the requirements, the first
time. For now, I'm just surprised that I accidentally got the right
answer to part 1 the first time, entirely by accident.