solve day17 part 1
I'm simultaneously really proud that this worked on the first go,
and proud of the "many small functions" nature of the code, and
kind of horrified all the same at the actual problem that I'm solving.

I went down a bit of a blind alley, really trying to leverage
existing Map traversal logic, but in the end it just wasn't a good fit.
I commend the authors for coming up with a problem that really does
require new code!