solve day23 part 2
Once again, an underwhelming part 2. Didn't have to do anything in
particular, just plug in the new value and run in release mode.
I almost wanted to change the computer to add a multiply instruction,
but decided that messing with the semantics without explicit
instruction would be dangerous.

Even part1 wasn't all that hard, though I'm not entirely happy with
the abstraction level.