solve day25
That was a nice high note to end on. The only point of confusion
today was that it turns out that Iterator.eq only matches when the
iterators have equal length as well as equal items, so I had to
`take` from both of them, not just one.

Assembunny was a good time. As for the rest of it, the difficulty
curve felt really weird this year; there were lots of easy ones
including very late, but occasionaly extreme difficulty spikes.

Not bad at all, though.